Compute contour gate on filtered data rather than raw dataset (as Olin's script).
NB: filtering is required in order to produce smooth gates. However, the density is influenced a lot by the filtering step in this case (ellipse and contour gates almost overlap).
Computing contour on the raw data is more accurate but produce a rugged contour. Anyhow gating on contour is not very appropriate since this might produce gates made of several connected components.